修改int和long错误
修改对象池错误

+++ TSFrame/Assets/TSFrame/Core/Model/ComponentValue.cs
@@ -1,7 +1,7 @@
 ﻿
 public struct ComponentValue
 {
-    public int ComponentId { get; set; }
+    public long ComponentId { get; set; }
 
     public string TSPropertyName { get; set; }
 }
